Add helper function "upsertConnection"
the upsertConnection function gives the ability to conditionally update a relationship.

For example, when updating what fuel card is connected to a vehicle, if the given fuel card id in the mutation is null, then disconnect, otherwise connect to the card with the inserted id.

Update README to reflect project name

Update package.json name value to reflect current project