Add code working towards use of refresh tokens
Currently authentication just uses a jwt token that is stored in the localStorafe of the browser. I would like to change this to use refresh tokens in tandem with the access tokens we are already using.

This will insure that the user does no have to relog in after 15 mins, while also making the authentication system more secure by way of storing the token in a http only token

(Not Complete)